requirements.txt
asgiref==3.6.0
Django==4.1.6
django_debug_toolbar==3.8.1
python-decouple==3.7
sqlparse==0.4.3
tzdata==2022.7
